requirements_teams.txt
flask==2.3.3
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
openai>=1.86.0,<2.0.0
azure-identity==1.15.0
python-dotenv==1.0.0
requests==2.31.0